+++ loc/src/Data/Loc/Types.hs
@@ -1,7 +1,6 @@
 -- | For convenience, this module exports only the important types from 'Data.Loc'
 module Data.Loc.Types
-  ( Pos,
-    Line,
+  ( Line,
     Column,
     Loc,
     Span,